kernel: printk: Allow logging to be toggled via sysfs
Taking a look at an article on LWN (https://lwn.net/Articles/705938/), I
have concluded that printk is something that can be problematic for
latency-sensitive systems such as Android. Printk grabs many locks and
can ultimately end in deadlocks, leaving it unsuitable for atomic
contexts (where it is often used). Allow printk to be disabled and
enabled during runtime to reduce potential overhead from printk.

The default state is disabled.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: clarencelol <clarencekuiek@icloud.com>
Signed-off-by: Anush02198 <Anush.4376@gmail.com>
Signed-off-by: clarencelol <clarencekuiek@icloud.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>